third_party/blink/web_tests/external/wpt/css/css-anchor-position/scrollable-containing-block-position-area-expected.txt

This is a testharness.js-based test.
[FAIL] .target 1
  assert_equals: \n<div class="target" style="position-area: top;" data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="70"></div>\nwidth expected 200 but got 120
[FAIL] .target 2
  assert_equals: \n<div class="target" style="position-area: right;" data-offset-x="120" data-offset-y="0" data-expected-width="80" data-expected-height="200"></div>\nwidth expected 80 but got 0
[FAIL] .target 3
  assert_equals: \n<div class="target" style="position-area: bottom;" data-offset-x="0" data-offset-y="120" data-expected-width="200" data-expected-height="80"></div>\nwidth expected 200 but got 120
[FAIL] .target 4
  assert_equals: \n<div class="target" style="position-area: left;" data-offset-x="0" data-offset-y="0" data-expected-width="70" data-expected-height="200"></div>\nheight expected 200 but got 120
Harness: the test ran to completion.
